Бот для чатов (3)/pyproject.toml
[tool.poetry]
name = "tg-knowledge-bot"
version = "0.1.0"
description = "Telegram knowledge bot: ASR/OCR/Docs → RAG summaries & search"
authors = ["You <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
aiogram = "^3.12.0"
pydantic = "^2.9.2"
orjson = "^3.10.7"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
loguru = "^0.7.2"
tenacity = "^9.0.0"
SQLAlchemy = "^2.0.36"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
redis = "^5.0.8"
rq = "^1.16.2"
ffmpeg-python = "^0.2.0"
Pillow = "^11.0.0"
pytesseract = "^0.3.13"
pypdf = "^5.1.0"
python-docx = "^1.1.2"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
unstructured = "^0.16.11"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"
faster-whisper = "^1.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
requests = "^2.32.3"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"